+++ src/Tests/WB.Tests.Unit/SharedKernels/Enumerator/ViewModels/CascadingSingleOptionQuestionViewModelTests/when_handling_SingleOptionQuestionAnswered_for_parent_question.cs
@@ -59,10 +59,10 @@ namespace WB.Tests.Unit.SharedKernels.Enumerator.ViewModels.CascadingSingleOptio
             cascadingModel.AutoCompleteSuggestions.Select(x => x.OriginalText).ShouldContainOnly(OptionsIfParentAnswerIs2.Select(x => x.Title));
 
         It should_create_option_models_with_specified_values = () =>
-            cascadingModel.AutoCompleteSuggestions.Select(x => Convert.ToInt64(x.Value)).ShouldContainOnly(OptionsIfParentAnswerIs2.Select(x => x.Value));
+            cascadingModel.AutoCompleteSuggestions.Select(x => Convert.ToInt32(x.Value)).ShouldContainOnly(OptionsIfParentAnswerIs2.Select(x => x.Value));
 
         It should_create_option_models_with_specified_ParentValues = () =>
-            cascadingModel.AutoCompleteSuggestions.Select(x => Convert.ToInt64(x.ParentValue)).ShouldContainOnly(OptionsIfParentAnswerIs2.Select(x => x.ParentValue.Value));
+            cascadingModel.AutoCompleteSuggestions.Select(x => Convert.ToInt32(x.ParentValue)).ShouldContainOnly(OptionsIfParentAnswerIs2.Select(x => x.ParentValue.Value));
 
         private static readonly List<CategoricalOption> OptionsIfParentAnswerIs2 = Options.Where(x => x.ParentValue == 2).ToList();
